Move logDirectoryProvide into server
