+++ extensions/typescript-language-features/src/typeScriptServiceClientHost.ts
@@ -21,7 +21,7 @@ import { CommandManager } from './utils/commandManager';
 import { Disposable } from './utils/dispose';
 import * as errorCodes from './utils/errorCodes';
 import { DiagnosticLanguage, LanguageDescription } from './utils/languageDescription';
-import { ILogDirectoryProvider } from './utils/logDirectoryProvider';
+import { ILogDirectoryProvider } from './tsServer/logDirectoryProvider';
 import { PluginManager } from './utils/plugins';
 import * as typeConverters from './utils/typeConverters';
 import TypingsStatus, { AtaProgressReporter } from './utils/typingsStatus';
